refactor(frontend): move parseApiError to $lib/api/
The parser is general-purpose API error handling — not claims-specific
and not editor-specific. Move parseApiError and FieldErrors to
$lib/api/parse-api-error.ts next to client.ts, with co-located tests
extracted from save-model-claims.test.ts. Drop the FieldErrors
re-export chain through save-claims-shared.ts and the four entity
save modules; consumers import directly from the canonical home.

Removes the corresponding task from the API boundary plan.
