+++ frontend/src/lib/components/editors/AliasesSectionEditor.svelte
@@ -3,7 +3,8 @@
   import TagInput from '$lib/components/form/TagInput.svelte';
   import type { SectionEditorProps } from './editor-contract';
   import { stringSetChanged } from '$lib/edit-helpers';
-  import type { FieldErrors, SaveMeta, SaveResult } from './save-claims-shared';
+  import type { FieldErrors } from '$lib/api/parse-api-error';
+  import type { SaveMeta, SaveResult } from './save-claims-shared';
 
   type AliasesData = { aliases: string[] };
