+++ frontend/src/lib/components/editors/DescriptionEditor.svelte
@@ -2,7 +2,8 @@
   import { untrack } from 'svelte';
   import MarkdownTextArea from '$lib/components/form/MarkdownTextArea.svelte';
   import type { SectionEditorProps } from './editor-contract';
-  import type { FieldErrors, SaveMeta, SaveResult } from './save-claims-shared';
+  import type { FieldErrors } from '$lib/api/parse-api-error';
+  import type { SaveMeta, SaveResult } from './save-claims-shared';
 
   type SaveFn = (
     slug: string,
