+++ frontend/src/lib/components/editors/NameEditor.svelte
@@ -7,7 +7,8 @@
   import type { SectionEditorProps } from './editor-contract';
   import { reconcileSlug, slugifyForCatalog } from '$lib/create-form';
   import { diffScalarFields, stringSetChanged } from '$lib/edit-helpers';
-  import type { FieldErrors, SaveMeta, SaveResult } from './save-claims-shared';
+  import type { FieldErrors } from '$lib/api/parse-api-error';
+  import type { SaveMeta, SaveResult } from './save-claims-shared';
 
   type NameFields = {
     name: string;
