+++ frontend/src/lib/components/editors/TitleExternalDataEditor.svelte
@@ -5,7 +5,8 @@
   import { fetchFieldConstraints, fc, type FieldConstraints } from '$lib/field-constraints';
   import { diffScalarFields } from '$lib/edit-helpers';
   import type { SectionEditorProps } from './editor-contract';
-  import { type FieldErrors, type SaveResult, type SaveMeta } from './save-claims-shared';
+  import type { FieldErrors } from '$lib/api/parse-api-error';
+  import type { SaveResult, SaveMeta } from './save-claims-shared';
   import { saveTitleClaims } from './save-title-claims';
 
   type ExternalDataTitle = {
