+++ frontend/src/routes/manufacturers/[slug]/edit/save-manufacturer-claims.ts
@@ -1,14 +1,10 @@
 import { invalidateAll } from '$app/navigation';
 import client from '$lib/api/client';
 import type { components } from '$lib/api/schema';
-import {
-  parseApiError,
-  type FieldErrors,
-  type SaveMeta,
-  type SaveResult,
-} from '$lib/components/editors/save-claims-shared';
+import { parseApiError } from '$lib/api/parse-api-error';
+import type { SaveMeta, SaveResult } from '$lib/components/editors/save-claims-shared';
 
-export type { FieldErrors, SaveMeta, SaveResult };
+export type { SaveMeta, SaveResult };
 
 type ManufacturerClaimsBody = components['schemas']['ClaimPatchSchema'];
